refactor(packaging): rename the gg_mcp_server import package to ggmcp
Align the top-level Python package with the distribution name. The module
tree moves from src/gg_mcp_server/ to src/ggmcp/ and all imports, the
console-script entrypoint (ggmcp.run:run_mcp_server), hatch package paths,
pyrefly scope, scripts, and tests reference the new name. No gg_mcp_server
references remain.
