+++ tests/test_server_composition.py
@@ -2,7 +2,7 @@
 
 from typing import Any
 
-from gg_mcp_server import server
+from ggmcp import server
 
 
 def test_get_server_initializes_observability_before_building_and_caches(monkeypatch):
